supersmartspeed/pyproject.toml
[tool.poetry]
name = "testingpermission"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
schedule = "^0.6.0"
speedtest-cli = "^2.1.2"
influxdb = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
